@@ -1,45 +1,46 @@
 import React from "react";
+import ReactSelect from "react-select";
 import InputRange from "react-input-range";
-import { useHistory } from "react-router-dom";
-import { Button, Form, InputGroup, ToggleButton, ToggleButtonGroup, Container, Row, Col } from "react-bootstrap";
+import { useNavigate } from "react-router-dom";
+import { Button, Form, ToggleButtonGroup, ToggleButton, Container, Row, Col } from "react-bootstrap";
 import { Editor } from 'ketcher-react'
 import { StandaloneStructServiceProvider } from 'ketcher-standalone'
 import { FontAwesomeIcon as Icon } from "@fortawesome/react-fontawesome";
 import { faSearch } from "@fortawesome/free-solid-svg-icons";
 
-import Select from "./Select.js"
 import { makeHash } from "./hash.js";
 import { databaseOpts, chargeOpts, isotopeOpts, aromaOpts, tautomerOpts, radicalOpts, stereoOpts, similarityOpts } from "./idsm.js" ;
 
 import "react-input-range/lib/css/index.css";
+import "ketcher-react/dist/index.css";
 
 
 function SelectOption(props) {
   return (
-    <Form.Row className="mt-1">
+    <Form.Group as={Row} className="mt-1">
       <Form.Label column xs={12} sm={4}>{props.label}</Form.Label>
       <Col xs={12} sm={8}>
-        <Select options={props.options} value={props.value} onChange={s => props.onChange(s)}/>
+        <ReactSelect options={props.options} value={props.value} onChange={s => props.onChange(s)}/>
       </Col>
-    </Form.Row>
+    </Form.Group>
   );
 }
 
 
 function CheckOption(props) {
   return (
-    <Form.Row className="mt-3">
+    <Form.Group as={Row} className="mt-3">
       <Col xs={12} sm={{ span: 8, offset: 4 }}>
-        <Form.Check type="checkbox" label={props.label} checked={props.value} onChange={t => props.onChange(t.target.checked)}/>
+        <Form.Check type="switch" label={props.label} checked={props.value} onChange={t => props.onChange(t.target.checked)}/>
       </Col>
-    </Form.Row>
+    </Form.Group>
   );
 }
 
 
 function RangeOption(props) {
   return (
-    <Form.Row className="mt-1">
+    <Form.Group as={Row} className="mt-1">
       <Form.Label column xs={12} sm={4}>{props.label}</Form.Label>
       <Col xs={12} sm={8}>
       <div style={{"margin-top":"0.8rem"}}>
@@ -48,7 +49,7 @@ function RangeOption(props) {
             formatLabel={(value) =>`${Math.round((value + Number.EPSILON) * props.scale) / props.scale}`}/>
       </div>
       </Col>
-    </Form.Row>
+    </Form.Group>
   );
 }
 
@@ -81,7 +82,7 @@ function SimilarityOptions(props) {
 
 
 function Search(props) {
-  const history = useHistory();
+  const navigate = useNavigate();
 
   const [state, setState] = React.useState(props.defaultParams);
 
@@ -105,34 +106,37 @@ function Search(props) {
         <Col xs={12} lg={8}><h1 className="display-4">Sachem GUI</h1></Col>
       </Row>
 
-      <Row className="mt-3">
-        <Col md={12} lg={7} xl={8} style={{ position: "inherit" }}>
-          <div id="ketcher" style={{ height:"600px" }}>
-            <Editor structServiceProvider={new StandaloneStructServiceProvider()}/>
-          </div>
+      <Row className="mt-3 row-gap-5">
+        <Col md={12} lg={7} xl={8}>
+          <Editor structServiceProvider={new StandaloneStructServiceProvider()} onInit={ketcher => {window.ketcher = ketcher}} />
         </Col>
 
         <Col md={6} lg={5} xl={4}>
           <Container>
             <h4>Search options</h4>
 
-            <ToggleButtonGroup type="radio" name="searchType" className="mt-2 w-100" defaultValue={state.mode} onChange={mode => {assignState({mode})}}>
-              <InputGroup.Prepend><InputGroup.Text>Search type</InputGroup.Text></InputGroup.Prepend>
-              <ToggleButton value={"substructure"} variant={state.mode === "substructure" ? "primary" : "outline-secondary"}>
-                Substructure
-              </ToggleButton>
-              <ToggleButton value={"similarity"} variant={state.mode === "similarity" ? "primary" : "outline-secondary"}>
-                Similarity
-              </ToggleButton>
-            </ToggleButtonGroup>
+            <Form.Group as={Row} className="mt-3 mb-3">
+              <Form.Label column xs={12} sm={4}>Search type</Form.Label>
+              <Col xs={12} sm={8}>
+                <ToggleButtonGroup type="radio" name="searchType" className="w-100" defaultValue={state.mode} onChange={mode => {assignState({mode})}}>
+                  <ToggleButton id="substructure-btn" value={"substructure"} variant={state.mode === "substructure" ? "primary" : "outline-secondary"}>
+                    Substructure
+                  </ToggleButton>
+                  <ToggleButton id="similarity-btn" value={"similarity"} variant={state.mode === "similarity" ? "primary" : "outline-secondary"}>
+                    Similarity
+                  </ToggleButton>
+                </ToggleButtonGroup>
+              </Col>
+            </Form.Group>
 
-            <Form.Row className="mt-3 mb-3">
+
+            <Form.Group as={Row} className="mt-3 mb-3">
               <Form.Label column xs={12} sm={4}>Database</Form.Label>
               <Col xs={12} sm={8}>
-                <Select isMulti closeMenuOnSelect={true} isClearable={false} options={databaseOpts}
+                <ReactSelect isMulti closeMenuOnSelect={true} isClearable={false} options={databaseOpts}
                   value={state.database} onChange={s => { if(s.length) assignState({ database: s }) }}/>
               </Col>
-            </Form.Row>
+            </Form.Group>
 
             {state.mode === "substructure" &&
               <SubstructureOptions values={state} onChange={(v) => assignState(v)}/>
@@ -142,19 +146,19 @@ function Search(props) {
               <SimilarityOptions values={state} onChange={(v) => assignState(v)}/>
             }
 
-            <Form.Row className="mt-4">
+            <Form.Group as={Row} className="mt-4">
               <Col xs={12} sm={{ span: 8, offset: 4 }}>
                 <Button variant="primary" size="lg" style={{ width: "100%" }} onClick={() => {
-                    window.ketcher.getMolfileAsync().then((molfile)=>{
+                    window.ketcher.getMolfile().then((molfile)=>{
                         const hash = makeHash(molfile, state);
-                        history.push("/search/" + hash);
-                        history.push("/results/" + hash);
+                        navigate("../search/" + hash);
+                        navigate("../results/" + hash);
                       });
                     }}>
                   <Icon icon={faSearch}/> Search
                 </Button>
               </Col>
-            </Form.Row>
+            </Form.Group>
           </Container>
         </Col>
       </Row>
