@@ -3,4 +3,4 @@ const servletBase = serverBase + "/chemweb";
 const endpointBase = serverBase;
 
 
-export { serverBase, servletBase, endpointBase };
+export { servletBase, endpointBase };
