@@ -29,7 +29,7 @@ function HomeCarousel() {
       <Carousel.Item style={{backgroundImage: `url(${idsmImage})`, backgroundSize: 'cover'}}>
         <Container fluid className="carousel-page d-flex flex-column justify-content-center text-white">
           <h1>Integrated Database of Small Molecules</h1>
-          <p className="text-justify">
+          <p>
             IDSM provides unique source of fast similarity and structural search functionality in databases such as
             ChEMBL, ChEBI or PubChem. The search interface is fully interoperable, and may be used in combination with
             other data sources using SPARQL.
@@ -49,7 +49,7 @@ function HomeCarousel() {
       <Carousel.Item style={{backgroundImage: `url(${sachemImage})`, backgroundSize: 'cover'}}>
         <Container fluid className="carousel-page d-flex flex-column justify-content-center text-white">
           <h1>Technology: Sachem</h1>
-          <p className="text-justify">
+          <p>
             Sachem is the open-source chemical cartridge that provides the high-performance indexing backend for the
             IDSM services. Installations of Sachem routinely handle hundreds of millions of stored compounds.
           </p>
@@ -68,7 +68,7 @@ function HomeCarousel() {
       <Carousel.Item style={{backgroundImage: `url(${rheaImage})`, backgroundSize: 'cover'}}>
         <Container fluid className="carousel-page d-flex flex-column justify-content-center text-white">
           <h1>Similarity search in Rhea reactome</h1>
-          <p className="text-justify">
+          <p>
             Interoperability allows you to search through Rhea, Uniprot, ChEMBL and IDSM data sources at once. We
             provide useful SPARQL query examples to get you started.
           </p>
@@ -91,7 +91,7 @@ function CardsRow() {
         <LinkContainer to="/sachem">
           <Button variant="link"><Image src={sachemIcon} width={140} height={140} roundedCircle/><h2>Sachem GUI</h2></Button>
         </LinkContainer>
-        <p className="text-justify flex-grow-1">
+        <p className="flex-grow-1">
           Sachem is a high-performance chemical cartridge for fingerprint-guided substructure and similarity search.
           Sachem GUI allows quick access to this functionality on our Sachem installation. Indexed databases include
           up-to-date versions of PubChem and ChEMBL.
@@ -105,7 +105,7 @@ function CardsRow() {
         <LinkContainer to="/sparql">
           <Button variant="link"><Image src={sparqlIcon} width={140} height={140} roundedCircle/><h2>SPARQL GUI</h2></Button>
         </LinkContainer>
-        <p className="text-justify flex-grow-1">
+        <p className="flex-grow-1">
           We provide an indexing service that allows FAIR-style search in published chemical data. You can easily
           construct queries that contain chemical substructure and similarity search terms combined with protein or
           bioassay-related queries.
@@ -118,7 +118,7 @@ function CardsRow() {
 
       <Col className="text-center d-flex flex-column mt-5">
         <Button href="/chemweb" variant="link"><Image src={chemwebIcon} width={140} height={140} roundedCircle/><h2>ChemWebRDF</h2></Button>
-        <p className="text-justify flex-grow-1">
+        <p className="flex-grow-1">
           PubChem RDF data exposed and searchable through an interoperable, semantic interface, using a custom
           high-performance SPARQL endpoint implementation. The service is still under development, but may already be
           used for processing many complicated queries.
@@ -138,7 +138,7 @@ function ChemistryRow() {
       </Col>
       <Col>
         <h2>Mission: Interoperability in small molecules</h2>
-        <p className="text-justify">
+        <p>
           The main goal of IDSM is to aggregate many different sources of information about small molecules into a
           single, logically coherent and semantically interconnected information source. This is achieved using the
           semantic web technologies. The data is explored and queried by the SPARQL query language; allowing to search
@@ -155,23 +155,23 @@ function DatabaseRow() {
     <Row className="align-items-center">
       <Col>
         <h2>IDSM technology</h2>
-        <p className="text-justify">
+        <p>
           IDSM internals have been described in several articles:
         </p>
         <ul>
-          <li className="text-justify">
+          <li>
             Galgonek, J. & Vondrášek, J.{' '}
             <b>IDSM ChemWebRDF: SPARQLing small-molecule datasets.</b>{' '}
             <em>Journal of Cheminformatics</em> <b>13</b>, 38 (2021).{' '}
             <a target="_blank" rel="noreferrer" href="https://doi.org/10.1186/s13321-021-00515-1">doi:10.1186/s13321-021-00515-1</a>
           </li>
-          <li className="text-justify">
+          <li>
             Kratochvíl, M., Vondrášek, J. & Galgonek, J.{' '}
             <b>Interoperable chemical structure search service.</b>{' '}
             <em>Journal of Cheminformatics</em> <b>11</b>, 45 (2019).{' '}
             <a target="_blank" rel="noreferrer" href="https://doi.org/10.1186/s13321-019-0367-2">doi:10.1186/s13321-019-0367-2</a>
           </li>
-          <li className="text-justify">
+          <li>
             Kratochvíl, M., Vondrášek, J. & Galgonek, J.{' '}
             <b>Sachem: a chemical cartridge for high-performance substructure search.</b>{' '}
             <em>Journal of Cheminformatics</em> <b>10</b>, 27 (2018).{' '}
