@@ -2,6 +2,7 @@ import React from "react";
 import { Container, Row, Col, Accordion, Card, Button } from "react-bootstrap";
 import { FontAwesomeIcon as Icon } from "@fortawesome/react-fontawesome";
 import { faBook, faInfoCircle, faServer, faPlay, faEdit } from "@fortawesome/free-solid-svg-icons";
+import { useAccordionButton } from 'react-bootstrap/AccordionButton';
 
 import { trackPageView } from "../matomo.js"
 
@@ -13,6 +14,17 @@ import { endpoints, defaultEndpoint, defaultQuery } from "./config";
 import { demoQueries } from "./demos";
 
 import "./Sparql.scss";
+
+
+function CustomToggle(props) {
+  const decoratedOnClick = useAccordionButton(props.eventKey);
+
+  return (
+    <Button variant={props.variant} className="w-100" onClick={decoratedOnClick}>
+      {props.children}
+    </Button>
+  );
+}
 
 
 function Sparql() {
@@ -31,7 +43,7 @@ function Sparql() {
         <Col xl={3} lg={4} md={5} sm={12} className="sparql-panel mb-4">
 
           <h2>IDSM SPARQL endpoints</h2>
-          <p className="text-justify">
+          <p>
             This page allows you to access the IDSM functionality through a SPARQL web-interface (on the right), and connect the search results to the data from other services. For a quick start, you may select a query example from the list below.
           </p>
           <p>
@@ -41,11 +53,11 @@ function Sparql() {
           <Accordion>
             <Card>
               <Card.Header>
-                <Accordion.Toggle as={Button} variant="outline-success" block eventKey={'0'}>
+                <CustomToggle variant="outline-success" block eventKey="0">
                   <Icon icon={faInfoCircle}/> Database status
-                </Accordion.Toggle>
+                </CustomToggle>
               </Card.Header>
-              <Accordion.Collapse eventKey={'0'}>
+              <Accordion.Collapse eventKey="0">
                 <Card.Body className="mt-2">
                   <Statistics url={servletBase + "/endpoint/statistics.json"}/>
                 </Card.Body>
@@ -56,11 +68,11 @@ function Sparql() {
           <Accordion onSelect={e => setStatusIsOpen(e != null)}>
             <Card>
               <Card.Header>
-                <Accordion.Toggle as={Button} variant="outline-success" block eventKey={'0'}>
+                <CustomToggle variant="outline-success" block eventKey="0">
                   <Icon icon={faServer}/> SPARQL endpoint status
-                </Accordion.Toggle>
+                </CustomToggle>
               </Card.Header>
-              <Accordion.Collapse eventKey={'0'}>
+              <Accordion.Collapse eventKey="0">
                 <Card.Body className="mt-2">
                   <Status endpoints={endpoints} isShown={statusIsOpen}/>
                 </Card.Body>
@@ -73,25 +85,25 @@ function Sparql() {
               demoQueries.map((group, index) =>
                 <Card key={'' + index}>
                   <Card.Header>
-                    <Accordion.Toggle as={Button} variant="outline-primary" block eventKey={'' + index}>
+                    <CustomToggle variant="outline-primary" block eventKey={'' + index}>
                       {group.name}
-                    </Accordion.Toggle>
+                    </CustomToggle>
                   </Card.Header>
                   <Accordion.Collapse eventKey={'' + index}>
                     <Card.Body className="demo-section">
-                      <p className="text-justify">{group.description}</p>
+                      <p>{group.description}</p>
                       {
                         group.queries.map((demo, subindex) =>
                           <Accordion key={'' + subindex}>
                             <Card>
                               <Card.Header>
-                                <Accordion.Toggle as={Button} variant="outline-secondary" block eventKey={'' + subindex}>
+                                <CustomToggle variant="outline-secondary" block eventKey={'' + subindex}>
                                   {demo.name}
-                                </Accordion.Toggle>
+                                </CustomToggle>
                               </Card.Header>
                               <Accordion.Collapse eventKey={'' + subindex}>
                                 <Card.Body>
-                                  <p className="text-justify">{demo.description}</p>
+                                  <p>{demo.description}</p>
                                   <Button onClick={() => {yasgui.current.demo(demo, true)}}><Icon icon={faPlay}/> Run demo</Button>
                                   {' '}
                                   <Button variant="secondary" onClick={() => {yasgui.current.demo(demo, false)}}><Icon icon={faEdit}/> Edit query</Button>
